// src/types/common.types.ts
export interface LocationAddressDetails {
  city: { longName: string; shortName: string };
  country: { longName: string; shortName: string };
  flatHouse: { longName: string; shortName: string };
  landmark: { longName: string; shortName: string };
  locality: { longName: string; shortName: string };
  state: { longName: string; shortName: string };
  street: { longName: string; shortName: string };
}